Refactor filename; use default export; add test
