+++ src/bricks/transformers/IdentityOptions.test.tsx
@@ -19,11 +19,12 @@ import React from "react";
 import { render } from "@/pageEditor/testHelpers";
 import IdentityOptions from "@/bricks/transformers/IdentityOptions";
 import { getExampleBrickConfig } from "@/pageEditor/exampleBrickConfigs";
-import { IdentityTransformer } from "@/bricks/transformers/identity";
+import IdentityTransformer from "@/bricks/transformers/IdentityTransformer";
 import brickRegistry from "@/bricks/registry";
 import { screen } from "@testing-library/react";
 import registerDefaultWidgets from "@/components/fields/schemaFields/widgets/registerDefaultWidgets";
 import registerEditors from "@/contrib/editors";
+import { toExpression } from "@/utils/expressionUtils";
 
 beforeAll(() => {
   brickRegistry.register([new IdentityTransformer()]);
@@ -32,6 +33,12 @@ beforeAll(() => {
 });
 
 describe("IdentityOptions", () => {
+  test("default config matches UI header", () => {
+    expect(getExampleBrickConfig(IdentityTransformer.BRICK_ID)).toEqual({
+      property: toExpression("nunjucks", ""),
+    });
+  });
+
   test("shows object widget by default", async () => {
     render(<IdentityOptions name="foo" configKey="config" />, {
       initialValues: {
